# specs/learning.openapi.yaml
...
openapi: 3.1.0
info:
  title: IT-Shaharcha — Learning Service
  version: 1.0.0
  description: IT skill courses, tutorials, reading docs, and typing practice.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: /api/v1
    description: Versioned platform base path (behind the gateway).
security:
  - bearerAuth: []
...
paths:
...
  /learning/courses/{courseId}:enroll:
    parameters:
      - name: courseId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: enrollCourse
      summary: Enroll the caller in a course. Emits learning.course.enrolled.
      tags: [enrollment]
      parameters:
        - $ref: "./shared/common.yaml#/components/parameters/IdempotencyKeyHeader"
      responses:
        "200":
          description: Enrollment (created or already present).
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "./shared/common.yaml#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Enrollment" }
        "401": { $ref: "./shared/common.yaml#/components/responses/Unauthorized" }
        "404": { $ref: "./shared/common.yaml#/components/responses/NotFound" }
  /learning/enrollments:
    get:
      operationId: listMyEnrollments
      summary: List the caller's enrollments with progress.
      tags: [enrollment]
      parameters:
        - $ref: "./shared/common.yaml#/components/parameters/PageParam"
        - $ref: "./shared/common.yaml#/components/parameters/SizeParam"
      responses:
        "200":
          description: A page of enrollments.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "./shared/common.yaml#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "./shared/common.yaml#/components/schemas/Page"
                          - type: object
                            properties:
                              items:
                                type: array
                                items: { $ref: "#/components/schemas/Enrollment" }
        "401": { $ref: "./shared/common.yaml#/components/responses/Unauthorized" }
  /learning/lessons/{lessonId}:start:
    parameters:
      - name: lessonId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: startLesson
      summary: Mark a lesson started. Emits learning.lesson.started.
      tags: [enrollment]
      responses:
        "204": { description: Lesson start recorded. }
        "401": { $ref: "./shared/common.yaml#/components/responses/Unauthorized" }
        "404": { $ref: "./shared/common.yaml#/components/responses/NotFound" }
  /learning/lessons/{lessonId}:complete:
    parameters:
      - name: lessonId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: completeLesson
      summary: Mark a lesson completed. Emits learning.lesson.completed.
      tags: [enrollment]
      parameters:
        - $ref: "./shared/common.yaml#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "./shared/events.yaml#/components/schemas/LessonCompletedData" }
      responses:
        "200":
          description: Progress updated.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "./shared/common.yaml#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/LessonProgress" }
        "401": { $ref: "./shared/common.yaml#/components/responses/Unauthorized" }
        "404": { $ref: "./shared/common.yaml#/components/responses/NotFound" }
        "422": { $ref: "./shared/common.yaml#/components/responses/ValidationFailed" }

  # ==========================================================================
  # Tutorials
  # ==========================================================================
  /learning/tutorials:
    get:
      operationId: listTutorials
      summary: List tutorial videos.
      tags: [tutorials]
      security: []
      parameters:
        - name: topic
          in: query
          required: false
          schema: { type: string }
        - $ref: "./shared/common.yaml#/components/parameters/PageParam"
        - $ref: "./shared/common.yaml#/components/parameters/SizeParam"
      responses:
        "200":
          description: A page of tutorials.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "./shared/common.yaml#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "./shared/common.yaml#/components/schemas/Page"
                          - type: object
                            properties:
                              items:
                                type: array
                                items: { $ref: "#/components/schemas/Tutorial" }
        "400": { $ref: "./shared/common.yaml#/components/responses/BadRequest" }
  /learning/tutorials/{tutorialId}:watched:
    parameters:
      - name: tutorialId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: recordTutorialWatched
      summary: Record tutorial watch progress. Emits learning.tutorial.watched.
      tags: [tutorials]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [watchedSeconds]
              properties:
                watchedSeconds: { type: integer, minimum: 0 }
                positionSeconds: { type: integer, minimum: 0 }
                completed: { type: boolean, default: false }
      responses:
        "204": { description: Watch progress recorded. }
        "401": { $ref: "./shared/common.yaml#/components/responses/Unauthorized" }
        "404": { $ref: "./shared/common.yaml#/components/responses/NotFound" }

  # ==========================================================================
  # Docs / reading
  # ==========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      $ref: "./shared/common.yaml#/components/securitySchemes/bearerAuth"
  schemas:
...
    Enrollment:
      type: object
      required: [id, courseId, accountId, status, enrolledAt]
      properties:
        id: { type: string, format: uuid, example: "77777777-7777-7777-7777-777777777777" }
        courseId: { type: string, format: uuid, example: "44444444-4444-4444-4444-444444444444" }
        accountId: { type: string, format: uuid, example: "11111111-1111-1111-1111-111111111111" }
        status:
          type: string
          enum: [active, completed, dropped]
          example: active
        progressPercent: { type: number, minimum: 0, maximum: 100, default: 0, example: 42.5 }
        enrolledAt: { type: string, format: date-time, example: "2026-05-20T08:00:00Z" }
        completedAt: { type: [string, "null"], format: date-time, example: null }
    LessonProgress:
      type: object
      required: [lessonId, courseId, accountId, completed]
      properties:
        lessonId: { type: string, format: uuid, example: "66666666-6666-6666-6666-666666666666" }
        courseId: { type: string, format: uuid, example: "44444444-4444-4444-4444-444444444444" }
        accountId: { type: string, format: uuid, example: "11111111-1111-1111-1111-111111111111" }
        completed: { type: boolean, example: true }
        scorePercent: { type: [number, "null"], minimum: 0, maximum: 100, example: 95 }
        courseProgressPercent: { type: number, minimum: 0, maximum: 100, example: 50 }
        completedAt: { type: [string, "null"], format: date-time, example: "2026-05-28T14:30:00Z" }
    Tutorial:
      type: object
      required: [id, title, videoUrl]
      properties:
        id: { type: string, format: uuid, example: "88888888-8888-8888-8888-888888888888" }
        title: { type: string, example: "Building a REST API in Node.js" }
        topic: { type: [string, "null"], example: "backend" }
        videoUrl: { type: string, format: uri, example: "https://videos.itsh.dev/rest-api-node.mp4" }
        durationSeconds: { type: [integer, "null"], minimum: 0, example: 1320 }
        thumbnailUrl: { type: [string, "null"], format: uri, example: "https://cdn.itsh.dev/thumbs/rest-api-node.jpg" }
...